backend/requirements.txt
--- Web Framework & Server ---
fastapi
uvicorn[standard]

--- Database ---
sqlalchemy
pymysql

--- Data Validation & Settings ---
pydantic
python-dotenv

--- Security (Passwords & JWT) ---
passlib[bcrypt]
python-jose[cryptography]

--- Google Gemini AI ---
google-generativeai